refactor: external(외부 api 호출) 패키지 디렉토리 리팩토링
