+++ external/src/main/java/com/laundreader/external/clova/dto/AssistantSuggestionEventDTO.java
@@ -1,11 +1,11 @@
-package com.laundreader.external.clova.service.response;
+package com.laundreader.external.clova.dto;
 
 import java.util.List;
 
 import lombok.Getter;
 
 @Getter
-public class AssistantSuggestionEvent {
+public class AssistantSuggestionEventDTO {
 	private String type;
 	private String message;
 	private List<String> suggestions;
