+++ external/src/main/java/com/laundreader/external/clova/dto/SummaryResultDTO.java
@@ -1,11 +1,11 @@
-package com.laundreader.external.clova.service.response;
+package com.laundreader.external.clova.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class SummaryResult {
+public class SummaryResultDTO {
 	private String text;
 	private int tokenCount;
 }
